CATEGORY
Created a js of default settings for css.
Fit to page view categories to use the values ConfigStylus.js
Created a new action creator in redux for categories, now when it in the page it does a fetch in the api and substitutes everything in the initial array. I did this because when I left and turned the page I would get and add in the array and the values were getting duplicated.